swift code form app to home for route
