+++ src/App.tsx
@@ -1,52 +1,12 @@
-import { FormEvent, useState } from 'react'
 import './App.css'
 import Greeting from './components/Greeting'
 import Navbar from './components/Navbar'
-import Post from './components/Post'
-import usePosts from './hooks/usePosts'
 
 function App() {
-  const { posts, isLoading, isSubmitting, createPost } = usePosts()
-  const [newTitle, setNewTitle] = useState<string>('')
-  const [newBody, setNewBody] = useState<string>('')
-
-  const handleSubmit = async (e: FormEvent) => {
-    e.preventDefault()
-
-    try {
-      await createPost(newTitle, newBody)
-
-      setNewTitle('')
-      setNewBody('')
-    } catch (err) {
-      console.error(err)
-    }
-  }
-
-  if (isLoading) return <h1>Loading...</h1>
-
   return (
     <div className="App">
       <Navbar />
       <Greeting name="Bun" isLoggedIn={true} />
-
-      <form onSubmit={handleSubmit}>
-        <label>Title</label>
-        <input type="text" value={newTitle} onChange={(e) => setNewTitle(e.target.value)} required />
-        <label>Body</label>
-        <input type="text" value={newBody} onChange={(e) => setNewBody(e.target.value)} required />
-
-        <button type="submit" disabled={isSubmitting}>
-          {isSubmitting ? 'Submitting...' : 'Submit'}
-        </button>
-      </form>
-
-      <div className="feed-container">
-        {posts &&
-          posts.map((post) => {
-            return <Post key={post.id} post={post} />
-          })}
-      </div>
     </div>
   )
 }
